2.ChatModels/chat_google.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model = ChatGoogleGenerativeAI(model='gemini-1.5-flash')

output = model.invoke('Who is the CEO of google?')

print(output.content)